requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.3.1
billiard==4.2.0
boto3==1.34.90
botocore==1.34.138
CacheControl==0.13.1
cachetools==5.3.3
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.2
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
cssselect2==0.7.0
daphne==4.1.2
Django==5.0.6
django-axes==6.4.0
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-countries==7.5.1
django-debug-toolbar==4.4.2
django-extensions==3.2.3
django-filter==23.2
django-ipware==7.0.1
django-jazzmin==3.0.0
django-js-asset==2.1.0
django-jsoneditor==0.2.4
django-modeltranslation==0.18.11
django-redis==5.4.0
django-silk==5.1.0
django-split-settings==1.3.1
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
environs==9.5.0
exceptiongroup==1.2.0
gprof2dot==2024.6.6
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
html5lib==1.1
httplib2==0.22.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.5
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.1
msgpack==1.0.5
oscrypto==1.3.0
packaging==23.1
Pillow==10.0.1
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.12.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.0.2
pypng==0.20220715.0
pytest==8.1.1
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-ipware==3.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
referencing==0.30.2
reportlab==4.0.9
requests==2.31.0
rpds-py==0.10.3
rsa==4.9
s3transfer==0.10.2
sentry-sdk==1.40.6
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.5.0